Move fork/money to common
