+++ app/components/common/money/index.js
