+++ app/components/common/money/money_bar.html
