+++ app/components/common/money/money_by_region.html
