+++ app/components/common/referendum/referendum_money.comp.js
@@ -3,7 +3,7 @@
 var angular = require('angular');
 
 angular.module('referendum.money', [
-  require('../fork/money'),
+  require('../money'),
   require('../array_update')
 ])
   .directive('referendumMoney', function () {
